HW Maven

+++ src/main/java/employee/Java/Employe/controller/EmployeeController.java
@@ -1,6 +1,8 @@
 package employee.Java.Employe.controller;
 import employee.Java.Employe.Employee;
 import employee.Java.Employe.EmployeeService;
+import employee.Java.Employe.Exception.EmployeeNotFoundException;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -18,11 +20,17 @@ public class EmployeeController {
 
     @GetMapping("/add")
     public Employee addEmployee (String firstName,String lastName,int department, int salary){
+        if (!StringUtils.isAlpha(firstName) || !StringUtils.isAlpha(lastName)) {
+            throw new EmployeeNotFoundException();
+        }
         return employeeService.add(firstName, lastName, department, salary);
     }
 
     @GetMapping("/remove")
     public Employee removeEmployee(String firstName, String lastName){
+        if (!StringUtils.isAlpha(firstName) || !StringUtils.isAlpha(lastName)) {
+            throw new EmployeeNotFoundException();
+        }
             return employeeService.remove(firstName, lastName);
     }
